Move types to their own file and add documentation
